backend/pyproject.toml
[project]
name = "stumbler-backend"
version = "0.1.0"
description = "Minimal FastAPI backend for Stumbler"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "wheel",
    "python-multipart>=0.0.20",
    "uvicorn>=0.30.6",
    "pydantic>2.9.2",
    "google-genai>=0.8.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "opencv-python",
    "mss",
]
